Pass global language options to packages
Both the `english` and `portuguese` options are now being passed to
babel via the document class.  The difference is the global option
defines which of the two is the main language.
Additionally, if `portuguese` is passed, it passes the `brazilian`
option to `cleveref` since they don't support a `portuguese` option.